refactor: create BubbleMenu component to split responsibilities between Editor and BubbleMenu
